requirements.txt
boto3==1.4.2
python-dateutil==2.6.0
